🔨 General code refactor
